Ajuste de detalles en codigo(cambiar los id a long) generar relaciones entre los modelos para facilitar el flujo de datos, realizo conexion a base de datos local en mysql

+++ src/main/java/cl/ecomarket/api/controller/tiendas/EmpleadoController.java
@@ -22,7 +22,7 @@ public class EmpleadoController {
         return empleadoService.guardarEmpleado(nuevoEmpleado);
     }
     @DeleteMapping("/{id}")
-    public void eliminarEmpleado(@PathVariable int id) {
+    public void eliminarEmpleado(@PathVariable Long id) {
         empleadoService.eliminarEmpleado(id);
     }
